// src/app/enums/responseEnums.ts
export enum responseEnums {
  SUCCESS = "SUCCESS",
  ERROR = "ERROR",
}
export enum exceptionEnums {
  SERVER_ERROR = "SERVER_ERROR",
  BAD_REQUEST = "BAD_REQUEST",
}
export enum userEnums {
  USER_EXISTS = "USER_EXISTS",
  INVALID_OTP = "INVALID_OTP",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INVALID_PASSWORD = "INVALID_PASSWORD",
  OTP_SUCCESS = "OTP_SUCCESS",
  ALREADY_LOGGED_IN = "ALREADY_LOGGED_IN",
  INVALID_USER = "INVALID_USER",
  INVALID_AMOUNT = "INVALID_AMOUNT",
  CONTACT_INFO_NOT_ALLOWED = "CONTACT_INFO_NOT_ALLOWED",
  UNAUTHORISED = "UNAUTHORISED",
}

export enum paymentEnums {
  PAYMENT_SUCCESS = "PAYMENT_SUCCESS",
  PAYMENT_FAILED = "PAYMENT_FAILED",
}